[CRITEO - Scala 2.10 Fix] Fixed non-serializable error in JavaMultilayerPerceptronClassifierSuite
I have failed to find the change between Scala 2.10 and 2.11
causing this, but operationally the issue originates from
'SparkSession.createDataFrame' which creates a (lazy!) 'Seq'
referencing a non-serializable 'java.util.ListIterator'.

I suspect one of the collection traits was change, but given how many
of them are in the standard library, it is hard to tell which one.
Here is a potential candidate scala/scala@0bb8a13