[CRITEO - Mesos fix] Fixed ALSCleanerSuite on Mesos
The test was overriding spark.local.dir with a temporary directory
and then checking that ALS produced the expected number of shuffle
files. Due to the implementation of Util.getConfiguredLocalDirs,
which prefers MESOS_DIRECTORY to spark.local.dir, the test failed
on Mesos.

This commits "solves" the problem in an and-hoc way by using the
result of Util.getConfiguredLocalDirs instead of a temporary directory.